Make Behaviors configurable.
You can now enable/disable behaviors you don't want using the bundle config.
By default, all behaviors are enabled. As soon as `knp_doctrine_behaviors` is
specified, all behaviors are disabled and the developer need to specify which
ones he want to enable.

This fix conflict issues, for example when a `setCurrentLocale` method is
defined by another third-party library/bundle.

This way does not introduce BC breaks, but only a new feature. The best would
be to add some interfaces and use them in subscribers. But that would imply
a major release... Maybe for the next one :-)