Shrink locale columns to 5 chars
Currently, the mapping for locale columns does not provide any length,
thus Doctrine falls back to the default size: 255 chars.

Not only such size is useless, but it also prevents using this library
with utf8mb4 charset, because InnoDB indexes are capped at 767 bytes,
and unfortunately: 4 * 255 = 1020 bytes.

For reference, here is the error with utf8mb4 charset:
> SQLSTATE[42000]: Syntax error or access violation: 1071 Specified key
> was too long; max key length is 767 bytes

Fixes #384.